feat(core) Support target position and quaternion offset in DistanceGrabbable.
Summary:
Added the support of target position and quaternion offset configuration in DistanceGrabbable. Support the target pos and quat calculation in Distance Handle and use lerp and slerp for the movement control.

Rev2: Add snap support when the object arrives at the input source and have it rest when the pointer is up.

Reviewed By: felixtrz

Differential Revision: D89073193

fbshipit-source-id: ea3867aee125d43a093ed250dd65ef538c554ad0